% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AML_BPCells}
\alias{AML_BPCells}
\title{Reference dataset used for testing and demonstration (Seurat with BPCells-backed matrices)}
\format{
An object of class \code{Seurat} with 462 rows and 250 columns.
}
\source{
The dataset was obtained from the \href{https://figshare.com/articles/dataset/Expression_of_197_surface_markers_and_462_mRNAs_in_15281_cells_from_blood_and_bone_marrow_from_a_young_healthy_donor/13398065/2}{Figshare}. For more information on the operations performed on the original object, see the \href{https://github.com/amc-heme/SCUBA_Manuscript/blob/main/Demo_Object_Generation.Rmd}{SCUBA manuscript repository}.
}
\usage{
AML_BPCells
}
\description{
A reference dataset for accute myeloid leukemia was included in this package for demonstration and testing. The data was originally published in \href{https://doi.org/10.1038/s41590-021-01059-0}{Triana et al. 2021}. The SCUBA authors downsampled the original Seurat object to use in the package for automated testing, and converted it into other object formats. The cell types provided by Triana et al. were also condensed into 10 generalized cell types to facilitate demonstration of SCUBA visualization capabilities. Details on the operations performed from the original object are provided in \href{https://github.com/amc-heme/SCUBA_Manuscript/blob/main/Demo_Object_Generation.Rmd}{this script} in the SCUBA manuscript repository.
}
\details{
This object is identical to \code{AML_Seurat}, but the counts and data matrices for the RNA and AB assays
are stored using \href{https://bnprks.github.io/BPCells/}{BPCells} on-disk matrices. The BPCells-backed
matrices enable efficient storage and access of large single-cell datasets while maintaining full
compatibility with SCUBA's data retrieval functions. The on-disk matrices are stored in the
\verb{inst/extdata/AML_BPCells/} directory.
}
\examples{
# Object summary
AML_BPCells

# Summary of metadata variables in object
meta_varnames(AML_BPCells)
}
\keyword{datasets}
