% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_names.R
\name{assay_names}
\alias{assay_names}
\alias{assay_names.Seurat}
\alias{assay_names.SingleCellExperiment}
\title{Get names of assays/experiments in object}
\usage{
assay_names(object, ...)

\method{assay_names}{Seurat}(object)

\method{assay_names}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{...}{Currently unused.}
}
\description{
Returns the names of all assays in a single-cell object. For Seurat objects,
returns assay names, and for SingleCellExperiment objects "experiments", the
equivalent of assays, are returned.
}
\section{Methods (by class)}{
\itemize{
\item \code{assay_names(Seurat)}: Seurat objects

\item \code{assay_names(SingleCellExperiment)}: SingleCellExperiment objects

}}
