% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_common_sections.R
\name{cells_param}
\alias{cells_param}
\title{Cells Parameter}
\arguments{
\item{cells}{A character vector of cell names to include, as they are named
in the object (i.e. according to \code{colnames(object)}). If \code{NULL}, data will
be returned for all cells in the object.}
}
\description{
Common parameter for functions that accept cell names.
}
\keyword{internal}
