% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pandas_index.R
\name{convert_pandas_index}
\alias{convert_pandas_index}
\title{Remove Pandas index attribute and set row names}
\usage{
convert_pandas_index(data)
}
\arguments{
\item{data}{a Python object returned via Reticulate operations that may be
a Pandas index.}
}
\value{
If \code{data} is a Pandas Index, it is converted to a list in Python
and returned as a character vector in R. Otherwise, it is returned as-is.
}
\description{
For MuData objects, queries such as object$obs_names will return a Pandas
Index, which will not be properly converted to a character vector when
accessed in R. \code{convert_pandas_index} converts a Pandas index to a Python
list so it can be properly converted to an R character vector by Reticulate.
}
\keyword{internal}
