% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_layer.R
\name{default_layer}
\alias{default_layer}
\alias{default_layer.Seurat}
\alias{default_layer.SingleCellExperiment}
\alias{default_layer.AnnDataR6}
\title{Get default reduction from object}
\usage{
default_layer(object, ...)

\method{default_layer}{Seurat}(object)

\method{default_layer}{SingleCellExperiment}(object)

\method{default_layer}{AnnDataR6}(object)
}
\arguments{
\item{object}{A single cell object. Currently, Seurat, SingleCellExpleriment,
and anndata objects are supported.}

\item{...}{Currently unused.}
}
\description{
Returns the default layer for the object passed. The default layer is chosen
based on the conventions used in each object to name the normalized counts
layer.
}
\section{Methods (by class)}{
\itemize{
\item \code{default_layer(Seurat)}: Seurat objects

For Seurat objects, the default layer is "data".

\item \code{default_layer(SingleCellExperiment)}: SingleCellExperiment objects

For SingleCellExperiment objects, the default layer is "logcounts".

\item \code{default_layer(AnnDataR6)}: Anndata objects

For Anndata objects, the default layer is \code{NULL}, which will direct
FetchData to pull feature epxression data from the X matrix.

}}
