% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_slot.R
\name{default_slot}
\alias{default_slot}
\alias{default_slot.Seurat}
\alias{default_slot.SingleCellExperiment}
\alias{default_slot.AnnDataR6}
\title{Get default reduction from object}
\usage{
default_slot(object, ...)

\method{default_slot}{Seurat}(object)

\method{default_slot}{SingleCellExperiment}(object)

\method{default_slot}{AnnDataR6}(object)
}
\arguments{
\item{object}{a single cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{...}{Currently unused.}
}
\description{
Returns the default slot for the object passed.
}
\section{Methods (by class)}{
\itemize{
\item \code{default_slot(Seurat)}: Seurat objects (default slot is "data")

\item \code{default_slot(SingleCellExperiment)}: SingleCellExperiment objects (default slot is "logcounts")

\item \code{default_slot(AnnDataR6)}: Anndata objects (default slot is NULL, to pull from X)

}}
