% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_plot.R
\name{expr_plot}
\alias{expr_plot}
\title{Plot feature expression by group}
\usage{
expr_plot(
  object,
  features,
  group_by,
  type = "violin",
  idents = NULL,
  ncol = NULL,
  sort = FALSE,
  y_max = NULL,
  same_y_lims = FALSE,
  adjust = 1,
  cols = NULL,
  pt_size = 0,
  split_by = NULL,
  log = FALSE,
  slot = NULL,
  stack = FALSE,
  combine = TRUE,
  fill_by = NULL,
  flip = FALSE,
  add_noise = TRUE,
  raster = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{group_by}{Group (color) cells in different ways (for example, orig.ident)}

\item{type}{Plot type, choose from 'ridge', 'violin', or 'splitViolin'}

\item{idents}{Which classes to include in the plot (default is all)}

\item{ncol}{Number of columns if multiple plots are displayed}

\item{sort}{Sort identity classes (on the x-axis) by the average expression
of the attribute being potted, or, if stack is True, sort both identity
classes and features by hierarchical clustering}

\item{y_max}{Maximum y axis value}

\item{same_y_lims}{Set all the y-axis limits to the same values}

\item{adjust}{Adjust parameter for geom_violin}

\item{cols}{Colors to use for plotting}

\item{pt_size}{Point size for geom_violin}

\item{split_by}{A variable to split the plot by}

\item{log}{plot Y axis on log scale}

\item{slot}{Slot (Seurat objects) or assay (SingleCellExperiment objects) to
pull expression data from (counts/data for Seurat objects, counts/logcounts
for SingleCellExperiment objects)}

\item{stack}{Horizontally stack plots for multiple feature}

\item{combine}{Combine plots into a single \code{\link[patchwork]{patchwork}ed}
ggplot object. If \code{FALSE}, return a list of ggplot objects}

\item{fill_by}{Color violins/ridges based on either 'feature' or 'ident'}

\item{flip}{flip plot orientation (identities on x-axis)}

\item{add_noise}{determine if adding a small noise for plotting}

\item{raster}{Convert points to raster format, default is \code{NULL} which
automatically rasterizes if plotting more than 100,000 cells}
}
\value{
A \code{\link[patchwork]{patchwork}ed} ggplot object if
\code{combine = TRUE}; otherwise, a list of ggplot objects
}
\description{
Common codebase for violin, ridge plots.
}
\keyword{internal}
