% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_data.R
\name{feature_data}
\alias{feature_data}
\title{Access expression data from a SingleCellExperiment object}
\usage{
feature_data(object, features, cells = NULL, slot = NULL)
}
\arguments{
\item{object}{a SingleCellExperiment object.}

\item{features}{a character vector with one or more features for which expression data will be fetched. To select features from experiments other than the main experiment, use the name of the desired alternate experiment followed by an underscore. For example, to fetch a surface protein from an alternate experiment named ADT, use \code{ADT_{name of protein}}.}

\item{cells}{a character vector giving the cells for which to fetch data. If \code{NULL}, data will be returned for all cells in the object.}

\item{slot}{the slot (assay in SingleCellExperiment objects) for which to pull data. To view the list of available assays in your object, use \code{assayNames({your object})}.}
}
\value{
A data.frame object containing expression data for the feature(s) entered.
}
\description{
This function is akin to the FetchData method for Seurat objects, and currently supports feature data only. Data from multiple assays (exeriments) can be selected by adding a prefix for the desired assay.
}
