% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_in_assay.R
\name{features_in_assay}
\alias{features_in_assay}
\alias{features_in_assay.Seurat}
\alias{features_in_assay.SingleCellExperiment}
\title{Get names of all features in an assay/experiment}
\usage{
features_in_assay(object, assay, ...)

\method{features_in_assay}{Seurat}(object, assay)

\method{features_in_assay}{SingleCellExperiment}(object, assay)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{assay}{the name of an assay/experiment for which to view features}

\item{...}{Currently unused.}
}
\description{
Returns the names of all features in an assay (Seurat objects) or experiment
(SingleCellExperiment objects)
}
\section{Methods (by class)}{
\itemize{
\item \code{features_in_assay(Seurat)}: Seurat objects

\item \code{features_in_assay(SingleCellExperiment)}: SingleCellExperiment objects

}}
