% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_common_sections.R
\name{features_param}
\alias{features_param}
\title{Features Parameter}
\arguments{
\item{features}{A character vector of feature names to retrieve from the
object.}
}
\description{
Common parameter for functions that accept feature names.
}
\keyword{internal}
