% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cells.R
\name{fetch_cells}
\alias{fetch_cells}
\alias{fetch_cells.Seurat}
\alias{fetch_cells.SingleCellExperiment}
\title{Subset cells based on metadata}
\usage{
fetch_cells(object, meta_var, meta_levels, ...)

\method{fetch_cells}{Seurat}(object, meta_var, meta_levels)

\method{fetch_cells}{SingleCellExperiment}(object, meta_var, meta_levels)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{meta_var}{a metadata variable used as the basis for subsetting cells.}

\item{meta_levels}{the levels of the specified metadata variable in
\code{meta_var} to include in the}

\item{...}{Currently unused.}
}
\description{
Returns a character vector with the cell names matching the levels/classes
of a specified metadata variable.
}
\section{Methods (by class)}{
\itemize{
\item \code{fetch_cells(Seurat)}: Seurat objects

\item \code{fetch_cells(SingleCellExperiment)}: SingleCellExperiment objects

}}
