% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_matrix.R
\name{fetch_matrix}
\alias{fetch_matrix}
\alias{fetch_matrix.AnnDataR6}
\title{Fetch matrix}
\usage{
fetch_matrix(object, matrix_location, ...)

\method{fetch_matrix}{AnnDataR6}(object, matrix_location, ...)
}
\arguments{
\item{object}{a single-cell object (Seurat, SingleCellExperiment, Anndata).}

\item{matrix_location}{the}

\item{...}{Currently unused.}
}
\description{
Returns object metadata for a specified set of cells.
}
\section{Methods (by class)}{
\itemize{
\item \code{fetch_matrix(AnnDataR6)}: AnnDataR6 objects

}}
\keyword{internal}
