% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_metadata.R
\name{fetch_metadata}
\alias{fetch_metadata}
\alias{fetch_metadata.Seurat}
\alias{fetch_metadata.SingleCellExperiment}
\title{Fetch metadata}
\usage{
fetch_metadata(object, vars, cells, return_class = "dataframe", ...)

\method{fetch_metadata}{Seurat}(object, vars, cells, return_class = "dataframe")

\method{fetch_metadata}{SingleCellExperiment}(object, vars, cells, return_class = "dataframe")
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{vars}{metadata variables to pull from object.}

\item{cells}{cells to include in the returned metadata.}

\item{return_class}{class of data returned. Set to "dataframe" by default to
return a data.frame, and may also be set to "vector" to yield a vector of
values.}

\item{...}{Currently unused.}
}
\description{
Returns object metadata for a specified set of cells.
}
\section{Methods (by class)}{
\itemize{
\item \code{fetch_metadata(Seurat)}: Seurat objects

\item \code{fetch_metadata(SingleCellExperiment)}: SingleCellExperiment objects

}}
