% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reduction.R
\name{fetch_reduction}
\alias{fetch_reduction}
\alias{fetch_reduction.Seurat}
\alias{fetch_reduction.SingleCellExperiment}
\alias{fetch_reduction.AnnDataR6}
\alias{fetch_reduction.BPCells}
\title{Fetch reduction coordinates from single-cell objects}
\usage{
fetch_reduction(object, reduction, cells = NULL, dims = c(1, 2))

\method{fetch_reduction}{Seurat}(object, reduction, cells = NULL, dims = c(1, 2))

\method{fetch_reduction}{SingleCellExperiment}(object, reduction, cells = NULL, dims = c(1, 2))

\method{fetch_reduction}{AnnDataR6}(object, reduction, cells = NULL, dims = c(1, 2))

\method{fetch_reduction}{BPCells}(object, reduction, cells = NULL, dims = c(1, 2))
}
\arguments{
\item{object}{A single cell object. Currently, Seurat, SingleCellExpleriment,
and anndata objects are supported.}

\item{reduction}{the name of the reduction to pull coordinates from.}

\item{cells}{cell IDs for which to pull reduction data. If \code{NULL},
coordinates will be returned from all cells in the object. Cell IDs can be
generated with \code{\link[=fetch_cells]{fetch_cells()}}.}

\item{dims}{a numeric vector indicating the dimensions to pull. Currently,
only two dimensions are supported, but \code{\link[=fetch_data]{fetch_data()}} supports more than
two dimensions. For instructions on pulling more than two dimensions at
once, see the examples of \code{\link[=fetch_data]{fetch_data()}}.}
}
\description{
Returns the reduction coordinates matching the name and dimensions supplied.
}
\details{
See our GitHub.io website for additional information and examples.
}
\section{Methods (by class)}{
\itemize{
\item \code{fetch_reduction(Seurat)}: Seurat objects

\item \code{fetch_reduction(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{fetch_reduction(AnnDataR6)}: AnnDataR6 objects

\item \code{fetch_reduction(BPCells)}: BPCells objects

}}
\examples{
# Return the first and second dimensions from the UMAP reduction
fetch_reduction(
  AML_Seurat,
  reduction = "umap",
  dims = c(1, 2)
  ) |> str()
}
