% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchData.R
\name{FetchData.SingleCellExperiment}
\alias{FetchData.SingleCellExperiment}
\title{FetchData Equivalent for SingleCellExperiment Objects}
\usage{
\method{FetchData}{SingleCellExperiment}(object, vars, layer = NULL, cells = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{vars}{A character vector with desired features or metadata variables
to pull from the object. To include features from an experiment other than
the main experiment, use the name of the experiment as a prefix (i.e. AB_CD4
for a feature in the experiment "AB" named "CD4".)}

\item{layer}{The assay (equivalent of layer in Seruat objects (slot in v4
and earlier)) to pull data from. To view the list of available assays in
your object, use \code{assayNames({your object})}.}

\item{cells}{A character vector of cells to include, as they are named in
the object (i.e. according to colNames(object)). If \code{NULL}, data will
be returned for all cells in the object.}
}
\value{
A data.frame object containing the requested expression
data or metadata.
}
\description{
The SingleCellExperiment equivalent of the SeuratObject FetchData method.
}
