% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_cells.R
\name{get_all_cells}
\alias{get_all_cells}
\alias{get_all_cells.Seurat}
\alias{get_all_cells.SingleCellExperiment}
\alias{get_all_cells.AnnDataR6}
\alias{get_all_cells.md._core.mudata.MuData}
\title{Get all cells from object}
\usage{
get_all_cells(object, ...)

\method{get_all_cells}{Seurat}(object)

\method{get_all_cells}{SingleCellExperiment}(object)

\method{get_all_cells}{AnnDataR6}(object)

\method{get_all_cells}{md._core.mudata.MuData}(object)
}
\arguments{
\item{object}{a single-cell object.}

\item{...}{Currently unused.}
}
\description{
Returns a character vector with all cell names in the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_all_cells(Seurat)}: Seurat objects

\item \code{get_all_cells(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{get_all_cells(AnnDataR6)}: SingleCellExperiment objects

\item \code{get_all_cells(md._core.mudata.MuData)}: MuData objects

}}
