% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_cells_utils.R
\name{get_bpcells_dir}
\alias{get_bpcells_dir}
\title{Get BPCells directory path}
\usage{
get_bpcells_dir(object, assay = NULL, layer = NULL)
}
\arguments{
\item{object}{A Seurat object containing BPCells-backed assay layers.}

\item{assay}{The name of the assay containing the BPCells-backed layer.
If \code{NULL}, the default assay will be used.}

\item{layer}{The name of the layer within the assay to query.
If \code{NULL}, the "data" layer will be used.}
}
\value{
A character string containing the directory path of the BPCells
matrix, or an error if the layer is not BPCells-backed.
}
\description{
Internal helper: retrieves the directory path for a BPCells-backed matrix
in a Seurat object. This function validates that the specified assay/layer
is BPCells-backed before attempting to access the directory path.
}
\keyword{internal}
