\name{scuba-internal}
\alias{chop.dive}
\alias{depths.dive}
\alias{is.dive}
\alias{is.gas}
\alias{is.rate}
\alias{nitroxname}
\alias{print.gas}
\alias{print.rate}
\alias{print.summary.dive}
\alias{print.summary.gas}
\alias{rate}
\alias{summary.dive}
\alias{summary.gas}
\alias{times.dive}
\alias{timetaken}
\title{Internal scuba functions}
\description{
  Internal scuba functions.
}
\usage{
chop.dive(d, tim)
depths.dive(d)
is.dive(x)
is.gas(x)
is.rate(x)
nitroxname(fO2)
\method{print}{gas}(x, \dots)
\method{print}{rate}(x, \dots)
\method{print}{summary.dive}(x, \dots)
\method{print}{summary.gas}(x, \dots)
rate(speed, time, up)
\method{summary}{dive}(object, \dots)
\method{summary}{gas}(object, \dots)
times.dive(d)
timetaken(start, finish, uprate, downrate)
}

\details{
  Not yet documented.
}
\keyword{internal}
