% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_type.R
\name{key_type}
\alias{key_type}
\alias{key_type.Seurat}
\alias{key_type.SingleCellExperiment}
\title{Get keys for all assays/reductions in an object}
\usage{
key_type(object, key, ...)

\method{key_type}{Seurat}(object, key)

\method{key_type}{SingleCellExperiment}(object, key)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{key}{the key to assess.}

\item{...}{Currently unused.}
}
\description{
Given a key, key_type will determine whether the key corresponds to an
assay/experiment, or a reduction.
}
\section{Methods (by class)}{
\itemize{
\item \code{key_type(Seurat)}: Seurat objects

\item \code{key_type(SingleCellExperiment)}: SingleCellExperiment objects

}}
