% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_common_sections.R
\name{layer_param}
\alias{layer_param}
\title{Layer Parameter}
\arguments{
\item{layer}{The layer to pull data from. Layers are referred to as "slots"
in Seurat objects v4 and earlier, and "assays" in SingleCellExperiment
objects. If \code{NULL}, the default layer will be used.}
}
\description{
Common parameter for functions that accept a layer name.
}
\keyword{internal}
