% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(file_path, backed = FALSE)
}
\arguments{
\item{file_path}{}

\item{backed}{}
}
\value{
obj
}
\description{
Load Data
}
\examples{
load_data("~/files/test.h5mu", backed="r+")
}
