% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_varnames.R
\name{meta_varnames}
\alias{meta_varnames}
\alias{meta_varnames.Seurat}
\alias{meta_varnames.SingleCellExperiment}
\alias{meta_varnames.AnnDataR6}
\title{Summarize object metadata}
\usage{
meta_varnames(object)

\method{meta_varnames}{Seurat}(object)

\method{meta_varnames}{SingleCellExperiment}(object)

\method{meta_varnames}{AnnDataR6}(object)
}
\arguments{
\item{object}{A single cell object. Currently, Seurat, SingleCellExpleriment,
and anndata objects are supported.}
}
\description{
Returns the names of all metadata variables in an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{meta_varnames(Seurat)}: Seurat objects

\item \code{meta_varnames(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{meta_varnames(AnnDataR6)}: Anndata objects

}}
\examples{
# Seurat objects
meta_varnames(AML_Seurat)

# SingleCellExperiment objects
meta_varnames(AML_SCE())

# anndata objects
meta_varnames(AML_h5ad())

}
