% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_common_sections.R
\name{object_param}
\alias{object_param}
\title{Object Parameter}
\arguments{
\item{object}{A single cell object. Currently, Seurat, SingleCellExpleriment,
and anndata objects are supported.}
}
\description{
The object parameter is used in nearly every instance of documentation in the SCUBA package. The parameter is described here to avoid excessive copy/pasting of documentation. When support is added for new object classes, the parameter description here should be updated.
}
\keyword{internal}
