% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridge.R
\name{plot_ridge}
\alias{plot_ridge}
\title{Single cell ridge plot}
\usage{
plot_ridge(
  object,
  features,
  cols = NULL,
  idents = NULL,
  sort = FALSE,
  assay = NULL,
  group_by = NULL,
  y_max = NULL,
  same_y_lims = FALSE,
  log = FALSE,
  ncol = NULL,
  slot = "data",
  stack = FALSE,
  combine = TRUE,
  fill_by = "feature"
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{cols}{Colors to use for plotting}

\item{idents}{Which classes to include in the plot (default is all)}

\item{sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted, can also pass 'increasing' or 'decreasing' to change sort direction}

\item{assay}{Name of assay to use, defaults to the active assay}

\item{group_by}{Group (color) cells in different ways (for example, orig.ident)}

\item{y_max}{Maximum y axis value}

\item{same_y_lims}{Set all the y-axis limits to the same values}

\item{log}{plot the feature axis on log scale}

\item{ncol}{Number of columns if multiple plots are displayed}

\item{slot}{Slot to pull expression data from (e.g. "counts" or "data")}

\item{stack}{Horizontally stack plots for each feature}

\item{combine}{Combine plots into a single \code{\link[patchwork]{patchwork}ed}
ggplot object. If \code{FALSE}, return a list of ggplot}

\item{fill_by}{Color violins/ridges based on either 'feature' or 'ident'}
}
\value{
A \code{\link[patchwork]{patchwork}ed} ggplot object if
\code{combine = TRUE}; otherwise, a list of ggplot objects
}
\description{
Draws a ridge plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
\examples{
data("pbmc_small")
RidgePlot(object = pbmc_small, features = 'PC_1')

}
\concept{visualization}
