% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot of single cell data}
\usage{
plot_scatter(
  object,
  feature_1,
  feature_2,
  group_by,
  cells = NULL,
  shuffle = FALSE,
  seed = 1,
  cols = NULL,
  pt_size = 1,
  shape_by = NULL,
  span = NULL,
  smooth = FALSE,
  combine = TRUE,
  slot = NULL,
  plot_cor = TRUE,
  raster = NULL,
  raster_dpi = c(512, 512),
  jitter = FALSE
)
}
\arguments{
\item{object}{a Seurat object or a SingleCellExperiment object}

\item{feature_1}{First feature to plot. Typically feature
expression but can also be QC metrics, PC scores, etc. - anything
that can be retrieved with FetchData}

\item{feature_2}{Second feature to plot.}

\item{group_by}{Name of one or more metadata variables to group
(color) cells by. Unlike \code{Seurat::FeatureScatter}, at least
one group_by variable must be defined}

\item{cells}{Cells to include on the scatter plot.}

\item{shuffle}{Whether to randomly shuffle the order of points. This can be
useful for crowded plots if points of interest are being buried. (default is FALSE)}

\item{seed}{Sets the seed if randomly shuffling the order of points.}

\item{cols}{Colors to use for identity class plotting.}

\item{pt_size}{Size of the points on the plot}

\item{shape_by}{Ignored for now}

\item{span}{Spline span in loess function call, if \code{NULL}, no spline added}

\item{smooth}{Smooth the graph (similar to smoothScatter)}

\item{combine}{Combine plots into a single \code{\link[patchwork]{patchwork}ed}}

\item{slot}{Slot to pull data from, should be one of 'counts', 'data', or 'scale.data'}

\item{plot_cor}{Display correlation in plot title}

\item{raster}{Convert points to raster format, default is \code{NULL}
which will automatically use raster if the number of points plotted is greater than
100,000}

\item{raster_dpi}{Pixel resolution for rasterized plots, passed to geom_scattermore().
Default is c(512, 512).}

\item{jitter}{Jitter for easier visualization of crowded points (default is FALSE)}
}
\value{
A ggplot object
}
\description{
Creates a scatter plot of two features (typically feature expression), across a
set of single cells. Cells are colored by their identity class. Pearson
correlation between the two features is displayed above the plot.
}
\keyword{internal}
