% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_violin}
\alias{plot_violin}
\title{Single cell violin plot}
\usage{
plot_violin(
  object,
  features,
  group_by,
  cols = NULL,
  pt_size = NULL,
  idents = NULL,
  sort = FALSE,
  assay = NULL,
  split_by = NULL,
  adjust = 1,
  y_max = NULL,
  same_y_lims = FALSE,
  log = FALSE,
  ncol = NULL,
  slot = NULL,
  split_plot = FALSE,
  stack = FALSE,
  combine = TRUE,
  fill_by = "feature",
  flip = FALSE,
  add_noise = TRUE,
  raster = NULL
)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object}

\item{features}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by fetch_data).}

\item{group_by}{Group (color) cells in different ways. Unlike
\code{Seurat::RidgePlot()} or \code{Seurat::VlnPlot()}, this must be defined
(SingleCellExperiment objects don't have \code{Idents()} functionality).}

\item{cols}{Colors to use for plotting}

\item{pt_size}{Point size for geom_violin}

\item{idents}{Which levels of the group by variable to include in the plot
(default is all)}

\item{sort}{Sort group by levels (on the x-axis) by the average
expression of the attribute being potted, can also pass 'increasing' or 'decreasing' to change sort direction}

\item{assay}{Name of assay to use, defaults to the active assay}

\item{split_by}{A variable to split the violin plots by,}

\item{adjust}{Adjust parameter for geom_violin}

\item{y_max}{Maximum y axis value}

\item{same_y_lims}{Set all the y-axis limits to the same values}

\item{log}{plot the feature axis on log scale}

\item{ncol}{Number of columns if multiple plots are displayed}

\item{slot}{Slot (Seurat objects) or assay (SingleCellExperiment objects) to
pull expression data from (counts/data for Seurat objects, counts/logcounts
for SingleCellExperiment objects)}

\item{split_plot}{plot each group of the split violin plots by multiple or
single violin shapes.}

\item{stack}{Horizontally stack plots for each feature}

\item{combine}{Combine plots into a single \code{\link[patchwork]{patchwork}ed}
ggplot object. If \code{FALSE}, return a list of ggplot}

\item{fill_by}{Color violins/ridges based on either 'feature' or 'ident'}

\item{flip}{flip plot orientation (identities on x-axis)}

\item{add_noise}{determine if adding a small noise for plotting}

\item{raster}{Convert points to raster format. Requires 'ggrastr' to be installed.}
}
\value{
A \code{\link[patchwork]{patchwork}ed} ggplot object if
\code{combine = TRUE}; otherwise, a list of ggplot objects
}
\description{
Draws a violin plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
\examples{
plot_violin(AML_Seurat, features = "PC_1", group_by = "condensed_cell_type")

}
\seealso{
\code{\link{FetchData}}
}
\concept{visualization}
\keyword{internal}
