% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_dimnames.R
\name{reduction_dimnames}
\alias{reduction_dimnames}
\alias{reduction_dimnames.Seurat}
\alias{reduction_dimnames.SingleCellExperiment}
\alias{reduction_dimnames.AnnDataR6}
\alias{reduction_dimnames.md._core.mudata.MuData}
\alias{reduction_dimnames.mudata._core.mudata.MuData}
\title{From names of reduction keys for fetch_data}
\usage{
reduction_dimnames(object, reduction, dims)

\method{reduction_dimnames}{Seurat}(object, reduction, dims)

\method{reduction_dimnames}{SingleCellExperiment}(object, reduction, dims)

\method{reduction_dimnames}{AnnDataR6}(object, reduction, dims)

\method{reduction_dimnames}{md._core.mudata.MuData}(object, reduction, dims)

\method{reduction_dimnames}{mudata._core.mudata.MuData}(object, reduction, dims)
}
\arguments{
\item{object}{A single cell object. Currently, Seurat, SingleCellExpleriment,
and anndata objects are supported.}

\item{reduction}{the name of the reduction.}

\item{dims}{a numeric vector with the dimensions for which
names should be returned.}
}
\description{
Given the name of a reduction and a set of dimensions, this function will
return the names of each dimension as it appears in the reduction matrix.
The output of this function can be passed to the \code{vars} parameter of
\code{\link[=fetch_data]{fetch_data()}} to facilitate the specification of reduction coordinates to
return from this function.
}
\section{Methods (by class)}{
\itemize{
\item \code{reduction_dimnames(Seurat)}: Seurat objects

\item \code{reduction_dimnames(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{reduction_dimnames(AnnDataR6)}: AnnDataR6 objects

\item \code{reduction_dimnames(md._core.mudata.MuData)}: MuData objects

\item \code{reduction_dimnames(mudata._core.mudata.MuData)}: MuData objects

}}
\examples{
# From names for first and second UMAP dimensions and 
# pass to fetch_data
dimnames <- reduction_dimnames(
  AML_Seurat,
  reduction = "umap",
  dims = c(1, 2)
  )
  
dimnames

fetch_data(
  AML_Seurat,
  vars = dimnames
  ) |> str()
  
  
# Form names for first five PCA dimensions and 
# pass to fetch_data
dimnames <- reduction_dimnames(
  AML_Seurat,
  reduction = "pca",
  dims = c(1:5)
  )
  
dimnames

fetch_data(
  AML_Seurat,
  vars = dimnames
  ) |> str()
}
