% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_names.R
\name{reduction_names}
\alias{reduction_names}
\alias{reduction_names.Seurat}
\alias{reduction_names.SingleCellExperiment}
\title{Get names of reductions in object}
\usage{
reduction_names(object, ...)

\method{reduction_names}{Seurat}(object)

\method{reduction_names}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{...}{Currently unused.}
}
\description{
Returns the names of all reductions in a single-cell object.
}
\section{Methods (by class)}{
\itemize{
\item \code{reduction_names(Seurat)}: Seurat objects

\item \code{reduction_names(SingleCellExperiment)}: SingleCellExperiment objects

}}
