% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_plot_relabel_axis.R
\name{relabel_axis}
\alias{relabel_axis}
\alias{relabel_axis.Seurat}
\alias{relabel_axis.SingleCellExperiment}
\alias{relabel_axis.AnnDataR6}
\title{Re-label axis behavior}
\usage{
relabel_axis(object, feature, ...)

\method{relabel_axis}{Seurat}(object, feature, ...)

\method{relabel_axis}{SingleCellExperiment}(object, feature, ...)

\method{relabel_axis}{AnnDataR6}(object, feature, ...)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat,
SingleCellExperiment, and Anndata objects are supported.}

\item{feature}{the feature being plotted. Only one feature should be passed
to this generic at once. For multi-feature plot, this generic should be ran
separately on each feature, and the changes in labeling applied to each
individual plot (patchwork object) iteratively.}

\item{...}{Currently unused.}
}
\description{
Code to relabel the axis in the expr_plot function. The default axis text is
"Feature Expression", and it changes or is removed based on the type of data
being plotted. If the data is a reduction, the label changes to "Embeddings
Value". If the data comes from a source that is not an assay or a reduction,
the label is removed. The label changed (x-axis or y-axis) depends on the
type of plot being created by expr_plot (violin, dot, etc.).
}
\details{
For Anndata objects, the label will always show as "Feature Expression",
unless the feature plotted is a metadata variable, in which case no label
will be drawn.

Labels can be manually added after plot creation with ggplot2::labs() in the
event a different label is more appropriate for the feature being plotted.
}
\section{Methods (by class)}{
\itemize{
\item \code{relabel_axis(Seurat)}: Seurat objects

\item \code{relabel_axis(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{relabel_axis(AnnDataR6)}: Anndata objects

}}
\keyword{internal}
