% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pandas_index.R
\name{remove_pandas_index}
\alias{remove_pandas_index}
\title{Remove Pandas index attribute and set row names}
\usage{
remove_pandas_index(table)
}
\arguments{
\item{table}{A \code{data.frame} object possibly containing a \code{pandas.index} attribute.}
}
\value{
Returns the modified \code{data.frame} with row names set
to the original Pandas index and the \code{pandas.index} attribute removed.
}
\description{
When Pandas Dataframes are converted to an R data.frame, they may still
contain a "pandas.index" attribute. In SCUBA, this happens for data extracted
via the fetch_* functions, which construct and return a Pandas DataFrame.
This function is used to fully convert data aggregated from Python object
formats to R syntax.
}
\keyword{internal}
