% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCUBA-package.R
\docType{package}
\name{SCUBA-package}
\alias{SCUBA}
\alias{SCUBA-package}
\title{SCUBA: Common data accession for single cell object formats}
\description{
SCUBA provides a common accession framework for multiple single cell object formats (Seurat, SingleCellExperiment, Anndata), facilitating analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://amc-heme.github.io/SCUBA/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
