\name{scuba.constants}
\alias{scuba.constants}
\title{Constants for Use in Scuba Package}
\description{
  A list of various useful physical constants
  needed in calculations in the scuba package.
}
\format{
  \code{scuba.constants} is a list containing the following entries:
  \tabular{ll}{
    \code{fsw.per.atm} \tab factor to convert atm (atmospheres absolute) \cr
                       \tab to fsw (feet of seawater) \cr
    \code{fO2air} \tab the fraction of oxygen in dry air at 25 C\cr
    \code{fN2air} \tab the fraction of \emph{nitrogen plus argon}
    in dry air at 25 C
  }
}
\author{Adrian Baddeley
  \email{Adrian.Baddeley@csiro.au}
  \url{http://www.maths.uwa.edu.au/~adrian/}
}
\keyword{utilities}
