\name{scuba.disclaimer}
\alias{scuba.disclaimer} %DoNotExport
\title{
  Disclaimer for Scuba Library
}
\description{
  This page contains important warnings.
} 
\details{
  The \code{scuba} software library is intended
  for use in research and education about the mathematical
  and statistical basis of decompression theory. It is
  not designed for actual use in scuba diving and related activities.
  It is emphatically not suitable for use in actual diving.

  Scuba diving is a dangerous activity with inherent risks of
  death and serious injury. No-one should attempt scuba diving
  without professional training, certification, supervision and
  regular medical assessment.

  It is also dangerous for trained scuba divers to exceed the
  limitations of their training. Diving at altitudes above sea level,
  and breathing mixed gases other than air, carry increased risk and
  additional types of risk. Divers should seek additional, professional
  training and certification for such activities. 

  This software is not suitable for use in actual scuba diving.
  The software will yield numerical results for any diving activity,
  without giving any warning if the activity would be dangerous or fatal.
  Each function in the \code{scuba} library calculates
  the predictions of one theoretical model (a law of physics,
  a decompression model or another empirical relationship).
  In doing so, it does not take account of safety restrictions,
  other physical laws, or other important information.

  The software is provided for academic interest only.
  It should not be used to generate diving tables or protocols
  related to diving. No output from this software should be misconstrued as
  a diving table. Only persons qualified to supervise diving
  activities or qualified in hyperbaric medicine should attempt to
  design diving tables. Although existing published diving tables are
  based on theoretical models, such tables have been extensively
  field-tested and modified before approval. Existing tables are more
  conservative than the models from which they were originally derived.

  The author does not warrant that the software is correct
  in any sense whatsoever. Even if correctly computed,
  the predictions of a theoretical physical model may not be correct
  predictions. 
}
\author{Adrian Baddeley
  \email{Adrian.Baddeley@uwa.edu.au}
  \url{http://www.maths.uwa.edu.au/~adrian/}
}
\keyword{utilities}
