% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pie.R
\name{shiny_pie_common}
\alias{shiny_pie_common}
\title{Common plotting actions across shiny_pie functions}
\usage{
shiny_pie_common(
  plot_data,
  group_by,
  palette = NULL,
  show_legend = TRUE,
  show_title = FALSE,
  plot_title = NULL
)
}
\arguments{
\item{plot_data}{object to be used for plotting.}

\item{group_by}{The metadata column to get patient count information for.
The pie chart will display the number of patients represented by each level
of this category.}

\item{palette}{The palette to use for coloring groups. If the palette passed
to this function is NULL, the default ggplot2 palette is used.}

\item{show_legend}{If TRUE, the legend is shown (default is TRUE)}

\item{show_title}{If TRUE, display a title above the plot. When TRUE, the
legend title is hidden, as the default legend title is equal to the default
plot title (the name of the group_by metadata column). The plot title can be
customized by supplying a string to \code{plot_title}.}

\item{plot_title}{If defined, the value entered will be displayed as the plot
title, if \code{show_title} is TRUE.}
}
\description{
shiny_pie_common() generates the plot and layer from incoming plot_data.
}
