% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_stacked_bar.R
\name{shiny_stacked_bar}
\alias{shiny_stacked_bar}
\title{Stacked bar plot}
\usage{
shiny_stacked_bar(
  object,
  group_by,
  split_by,
  x_axis_title = NULL,
  y_axis_title = NULL,
  plot_title = NULL,
  show_title = TRUE,
  show_legend = TRUE,
  palette = NULL,
  sort_groups = NULL,
  custom_factor_levels = NULL,
  legend_ncol = NULL,
  legend_font_size = NULL,
  legend_key_size = NULL
)
}
\arguments{
\item{object}{A single cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{group_by}{metadata variable used for cell type proportions. The number
of cells in each level of this variable will be plotted.}

\item{split_by}{metadata variable used for comparing cell type proportions.
The number of cells in each level of this variable will be plotted.}

\item{x_axis_title}{the title to use for the x-axis.}

\item{y_axis_title}{the title to use for the y-axis.}

\item{plot_title}{the title of the plot. Defaults to NULL; in this case, the group_by variable name will be displayed.}

\item{show_title}{if TRUE, the title is displayed above the plot.}

\item{show_legend}{if TRUE, the legend is shown to the right side of the plot. The default is TRUE.}

\item{palette}{a color palette to use for the plot. The plot uses a categorical palette.}

\item{sort_groups}{the order with which to sort proportion comparisons on the proportion plot. This may be set to "ascending" or "descending". If ascending, groups will be sorted in increasing alphabetical order. If descending, they will be sorted in decreasing alphabetical order.}

\item{custom_factor_levels}{A character vector giving the order of groups if \code{sort_groups} is set to "custom".}
}
\value{
a ggplot2 object with a stacked bar plot created according to user specifications.
}
\description{
Creates a stacked bar plot showing the proportion of cells belonging to each
level of a metadata variable. One bar is created for each level of a second
metadata variable, making the chart able to compare metadata proportions
across the second variable. For example, the proportion of cells belonging
to each cluster in the object can be compared across different samples, with
one bar shown for each sample. The bar will consist of a series of stacked
bars, the heights of which being proportional to the percentage of cells in
the sample belonging to each cluster.
}
