% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_metadata.R
\name{unique_metadata}
\alias{unique_metadata}
\title{Unique metadata values}
\usage{
unique_metadata(object, var)
}
\arguments{
\item{object}{any object type supported by
\code{\link[fetch_metadata]{fetch_metadata}}.}

\item{var}{the metadata variable for which to return unique values.}
}
\value{
a character vector giving the unique values in the specified
metadata variable.
}
\description{
Returns the unique values for a given metadata variable in a single cell object.
The function supports all object types that can be passed \code{\link[fetch_metadata]{fetch_metadata}}.
}
