% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_values.R
\name{unique_values}
\alias{unique_values}
\title{Unique metadata values}
\usage{
unique_values(object, var)
}
\arguments{
\item{object}{any object type supported by \code{\link[=fetch_metadata]{fetch_metadata()}}.}

\item{var}{the metadata variable for which to return unique values.}
}
\value{
a character vector giving the unique values in the specified
metadata variable.
}
\description{
Returns the unique values for a given metadata variable in a single cell object.
The function supports all object types that can be passed \code{\link[=fetch_metadata]{fetch_metadata()}}.
}
