% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_keys.R
\name{all_keys}
\alias{all_keys}
\alias{all_keys.Seurat}
\alias{all_keys.SingleCellExperiment}
\alias{all_keys.AnnDataR6}
\alias{all_keys.md._core.mudata.MuData}
\title{Get keys for all assays/reductions in an object}
\usage{
all_keys(object, ...)

\method{all_keys}{Seurat}(object)

\method{all_keys}{SingleCellExperiment}(object)

\method{all_keys}{AnnDataR6}(object)

\method{all_keys}{md._core.mudata.MuData}(object)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{...}{Currently unused.}
}
\description{
Returns all keys in the object. For SingleCellExperiment objects, currently
returns only the names of all reductions in the object, as well as all
experiments.
}
\section{Methods (by class)}{
\itemize{
\item \code{all_keys(Seurat)}: Seurat objects

\item \code{all_keys(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{all_keys(AnnDataR6)}: SingleCellExperiment objects

\item \code{all_keys(md._core.mudata.MuData)}: MuData objects

}}
