% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AML_h5ad}
\alias{AML_h5ad}
\title{Reference dataset used for testing and demonstration (anndata)}
\source{
\url{https://figshare.com/articles/dataset/Expression_of_197_surface_markers_and_462_mRNAs_in_15281_cells_from_blood_and_bone_marrow_from_a_young_healthy_donor/13398065/2}
}
\usage{
AML_h5ad()
}
\description{
A reference dataset for accute myeloid leukemia was included in this package for demonstration and testing. The data was originally published in \href{https://doi.org/10.1038/s41590-021-01059-0}{Triana et al. 2021}. The SCUBA authors downsampled the original Seurat object to use in the package for automated testing, and converted it into other object formats. The cell types provided by Triana et al. were also condensed into 10 generalized cell types to facilitate demonstration of SCUBA visualization capabilities. Details on the operations performed from the original object, downloaded from \href{https://figshare.com/articles/dataset/Expression_of_197_surface_markers_and_462_mRNAs_in_15281_cells_from_blood_and_bone_marrow_from_a_young_healthy_donor/13398065/2}{figshare}, are provided in \href{https://github.com/amc-heme/SCUBA_Manuscript/blob/main/Demo_Object_Generation.Rmd}{this script} in the SCUBA manuscript repository.
}
\details{
Contrary to convention for loading data, \code{AML_h5ad()} is called as a function, with parentheses. This is because the dataset is in Python, and can't be loaded using the typical process of loading data included with R packages.
}
\examples{
# These examples require a functional Python installation with prerequisite
# packages installed to work
# Please see our website for more details
# https://amc-heme.github.io/SCUBA/index.html#installation
# The examples may take a while to run the first time they are executed in 
# a session due to the time required to initialize a Python environment. 
# R Studio does not display a spinner while these run, so the "run_examples"
# link may not appear to do anything until the examples are finished. 
AML_h5ad()

# Summary of metadata variables in object
meta_varnames(AML_h5ad())
}
