% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_layer.R
\name{default_layer}
\alias{default_layer}
\alias{default_layer.Seurat}
\alias{default_layer.SingleCellExperiment}
\alias{default_layer.AnnDataR6}
\alias{default_layer.md._core.mudata.MuData}
\title{Return the default layer}
\usage{
default_layer(object)

\method{default_layer}{Seurat}(object)

\method{default_layer}{SingleCellExperiment}(object)

\method{default_layer}{AnnDataR6}(object)

\method{default_layer}{md._core.mudata.MuData}(object)
}
\arguments{
\item{object}{A single cell object. Currently, Seurat, SingleCellExpleriment,
and anndata objects are supported.}
}
\description{
Returns the default layer for the object passed. The default layer is chosen
based on the conventions used in each object to name the normalized counts
layer.
}
\details{
This is a utility function that is most useful for defining defaults in plotting functions, to reduce the number of required parameters end users need to supply. \url{see our website} for an example of usage in a function.
}
\section{Methods (by class)}{
\itemize{
\item \code{default_layer(Seurat)}: Seurat objects

For Seurat objects, the default layer is "data".

\item \code{default_layer(SingleCellExperiment)}: SingleCellExperiment objects

For SingleCellExperiment objects, the default layer is "logcounts".

\item \code{default_layer(AnnDataR6)}: Anndata objects

For Anndata objects, the default layer is \code{NULL}, which will direct
FetchData to pull feature expression data from the X matrix.

\item \code{default_layer(md._core.mudata.MuData)}: MuData objects

Default layer is \code{NULL} for MuData objects, which will direct
FetchData to pull feature expression data from the X matrix of the
anndata object representing each modality.

}}
\examples{
# Seurat objects
default_layer(AML_Seurat)

# SingleCellExperiment objects
default_layer(AML_SCE())

# anndata objects
default_layer(AML_h5ad())
}
