% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_reduction.R
\name{default_reduction}
\alias{default_reduction}
\alias{default_reduction.Seurat}
\alias{default_reduction.SingleCellExperiment}
\alias{default_reduction.AnnDataR6}
\title{Get default reduction from object}
\usage{
default_reduction(object, ...)

\method{default_reduction}{Seurat}(object)

\method{default_reduction}{SingleCellExperiment}(object)

\method{default_reduction}{AnnDataR6}(object)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{...}{Currently unused.}
}
\description{
Returns a reduction from the object to use as the default.
}
\section{Methods (by class)}{
\itemize{
\item \code{default_reduction(Seurat)}: Seurat objects

\item \code{default_reduction(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{default_reduction(AnnDataR6)}: AnnDataR6 objects

}}
