% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defult_slot.R
\name{default_slot}
\alias{default_slot}
\title{Get default reduction from object}
\usage{
default_slot(object)
}
\arguments{
\item{object}{a single cell object supported by SCUBA. Currently, Seurat, SingleCellExperiment, and anndata objects are supported.}
}
\description{
Returns the default slot for the object passed. This function was deprecated
in version 0.10.0. Please use \code{\link[=default_layer]{default_layer()}} instead.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
