% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_metadata.R
\name{fetch_metadata}
\alias{fetch_metadata}
\alias{fetch_metadata.Seurat}
\alias{fetch_metadata.SingleCellExperiment}
\title{Fetch metadata}
\usage{
fetch_metadata(object, vars, cells, ...)

\method{fetch_metadata}{Seurat}(object, vars, cells)

\method{fetch_metadata}{SingleCellExperiment}(object, vars, cells)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{vars}{metadata variables to pull from object}

\item{cells}{cells to include in the returned metadata}

\item{...}{Currently unused.}
}
\description{
Returns object metadata for a specified set of cells.
}
\section{Methods (by class)}{
\itemize{
\item \code{fetch_metadata(Seurat)}: Seurat objects

\item \code{fetch_metadata(SingleCellExperiment)}: SingleCellExperiment objects

}}
