% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reduction.R
\name{fetch_reduction}
\alias{fetch_reduction}
\alias{fetch_reduction.Seurat}
\alias{fetch_reduction.SingleCellExperiment}
\title{Get names of reduction keys}
\usage{
fetch_reduction(object, reduction, dims, ...)

\method{fetch_reduction}{Seurat}(object, reduction, cells, dims)

\method{fetch_reduction}{SingleCellExperiment}(object, reduction, cells, dims)
}
\arguments{
\item{object}{a single-cell object. Currently, Seurat and
SingleCellExperiment objects are supported.}

\item{reduction}{the reduction from which names should be formed}

\item{dims}{a two-element integer vector with the dimensions for which
names should be returned}

\item{...}{Currently unused.}
}
\description{
Given the name of a reduction and a set of dims, this function will return
the corresponding reduction data.
}
\section{Methods (by class)}{
\itemize{
\item \code{fetch_reduction(Seurat)}: Seurat objects

\item \code{fetch_reduction(SingleCellExperiment)}: SingleCellExperiment objects

}}
