% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_data.R
\name{FetchData.SingleCellExperiment}
\alias{FetchData.SingleCellExperiment}
\title{FetchData Equivalent for SingleCellExperiment Objects}
\usage{
\method{FetchData}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\description{
FetchData Equivalent for SingleCellExperiment Objects
}
