% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_cells.R
\name{get_all_cells}
\alias{get_all_cells}
\alias{get_all_cells.Seurat}
\alias{get_all_cells.SingleCellExperiment}
\alias{get_all_cells.AnnDataR6}
\title{Plotting utility function to return all cell IDs}
\usage{
get_all_cells(object)

\method{get_all_cells}{Seurat}(object)

\method{get_all_cells}{SingleCellExperiment}(object)

\method{get_all_cells}{AnnDataR6}(object)
}
\arguments{
\item{object}{A single cell object. Currently, Seurat, SingleCellExpleriment,
and anndata objects are supported.}
}
\description{
Returns a character vector with all cell names in the object. This is a utility function used to set defaults for plotting functions created by SCUBA
}
\details{
For additional information, see our GitHub.io website ("User Guide" article, "Data Visualization" section)
}
\section{Methods (by class)}{
\itemize{
\item \code{get_all_cells(Seurat)}: Seurat objects

\item \code{get_all_cells(SingleCellExperiment)}: SingleCellExperiment objects

\item \code{get_all_cells(AnnDataR6)}: SingleCellExperiment objects

}}
\examples{
get_all_cells(AML_Seurat) |> str()
}
