\name{scuba-internal}
\alias{chop.dive}
\alias{depths.dive}
\alias{is.dive}
\alias{is.gas}
\alias{is.rate}
\alias{print.gas}
\alias{print.rate}
\alias{print.summary.dive}
\alias{print.summary.gas}
\alias{rate}
\alias{summary.dive}
\alias{summary.gas}
\alias{times.dive}
\alias{timetaken}
\title{Internal scuba functions}
\description{
  Internal scuba functions.
}
\usage{
chop.dive(d, tim)
depths.dive(d)
is.dive(x)
is.gas(x)
is.rate(x)
print.gas(x, \dots)
print.rate(x, \dots)
print.summary.dive(x, \dots)
print.summary.gas(x, \dots)
rate(speed, time, up)
summary.dive(object, \dots)
summary.gas(object, \dots)
times.dive(d)
timetaken(start, finish, uprate, downrate)
}

\details{
  Not yet documented.
}
\keyword{internal}
