% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_unique.R
\name{n_unique}
\alias{n_unique}
\title{Object metadata: number of unique values}
\usage{
n_unique(object, meta_var)
}
\arguments{
\item{object}{A single cell object.}

\item{meta_var}{Name of a metadata variable.}
}
\value{
Number of unique classes (integer).
}
\description{
Returns the number of unique values in a single cell object, for the specified
metadata variable. Any object type supported by
\code{\link[unique_values]{unique_values}} may be used.
}
\keyword{internal}
