% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCUBA-package.R
\docType{package}
\name{SCUBA-package}
\alias{SCUBA}
\alias{SCUBA-package}
\title{SCUBA: Common data accession for single cell object formats}
\description{
SCUBA provides a common accession framework for multiple single cell object formats (Seurat, SingleCellExperiment, Anndata), facilitating analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://amc-heme.github.io/SCUBA/}
}

}
\author{
\strong{Maintainer}: Austin Gillen \email{austin.gillen@cuanschutz.edu} [funder]

Authors:
\itemize{
  \item William Showers
  \item Jairav Desai
}

Other contributors:
\itemize{
  \item Kent Riemondy [contributor]
  \item Krysta Engel [contributor]
  \item Clayton Smith [funder]
  \item Craig Jordan [funder]
}

}
\keyword{internal}
