% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_cells_utils.R
\name{set_bpcells_dir}
\alias{set_bpcells_dir}
\title{Set BPCells directory path}
\usage{
set_bpcells_dir(object, assay, layer, dirname)
}
\arguments{
\item{object}{A Seurat object containing BPCells-backed assay layers.}

\item{assay}{The name of the assay containing the BPCells-backed layer.}

\item{layer}{The name of the layer within the assay to modify.}

\item{dirname}{The target directory path to set for the BPCells matrix.
Can be specified as either an absolute or relative path; the path will
be normalized to an absolute path before storage.}
}
\value{
The modified Seurat object (invisibly). The object is returned
invisibly so it will not print when called interactively. To apply
the changes, you must assign the output back to the object:
\code{obj <- set_bpcells_dir(obj, "RNA", "data", "/path")}.
Alternatively, use the pipe operator:
\code{obj <- obj |> set_bpcells_dir("RNA", "data", "/path")}.
}
\description{
Internal helper: sets the directory path for a BPCells-backed matrix in a
Seurat object. This function validates that the specified assay/layer is
BPCells-backed before attempting to modify the directory path.
}
\keyword{internal}
